RemoveUnusedBrs: Refinalize when removing a break (#7918)
We refinalized when changing a break's type, but not when removing
it entirely. But that removal can affect other types, so we need to.

Looks like a regression from #7639, which was very hard for the fuzzer
to find. The issue is that we end up with types not fully refinalized, but
on control flow, that the validator doesn't fully check. The concrete
issue was some weird sequence of passes that leads to an assertion
inside SimplifyLocals; I extracted a simpler testcase using debug
annotations.