DeadArgumentElimination: Refine result types of exported functions (#7935)
Exports prevent most optimizations in this pass (we can't remove a
parameter, for example), but we can refine results: the export will just
return more-refined things, which is fine.

This also helps the fuzzer, in that it allows us to generate more
interesting testcases - we can modify the types of exports after the
fact.