Optimize exact casts better in Unsubtyping (#7930)
Casts generally need special treatment in Unsubtyping because any
subtype of the destination type could flow into the source type and be
successfully cast, meaning all original subtypes of the destination type
that are still subtypes of the source types need to remain subtypes of
the destination type as well. However, exact casts do not have this
property because strict subtypes of the destination type will fail the
cast whether or not they remain subtypes of the destination type in the
optimized module.

Update the analysis so that exact casts require only simple subtyping
between the destination and source types.