Revert "DeadArgumentElimination: Refine result types of exported functions" (#7943)
Reverts WebAssembly/binaryen#7935

Refining exports is unsafe given exact casts: the importing module
can notice that we refined, as exact casts to the old type will fail.

Doing this only when custom descriptors (exact casts) are disabled
is possible, but rarely useful, and it's better to focus on
optimizations
that work with newer features.