[Custom Descriptors][NFC] Simplify TypeMerging (#7946)
Now that we have stricter validation for descriptor types, descriptor
chains can only be subtypes of other descriptor chains if the base
described type of the subtype chain is a subtype of the base described
type of the supertype chain. We had previously handled more complicated
cases where the base described type in the supertype chain could be a
supertype of an arbitrary type in the subtype chain and where a subtype
chain could even have several supertype chains.

Simplify TypeMerging by reverting #7842 and #7764 now that we no longer
need to handle these more complicated cases.