[NFC] Model imported and exported function values
Update ExternalInterface to provide imported functions as values
rather than just exposing an interface to call them. Similarly, update
a ModuleRunner instance to provide exported functions as values rather
than just providing an interface to call them.

Dealing with imported and exported functions as values like this is a
prerequisite for fixing our interpretation of function reference casts
to reflect that fact that the runtime type of a reference to an imported
function may be a nontrivial subtype of the type of the corresponding
function import. Actually fixing this is left to a follow-on PR.