Fuzzer: Add a flag to import another module in --translate-to-fuzz (#7949)
-ttf --fuzz-import=foo.wasm will generate a fuzz new file, and that file
will import parts of the given foo.wasm. So far, this imports a subset
of the functions of that other module.

Use this in the Two fuzzer as well as in ClusterFuzz.

Add fuzz_shell.js support for importing a second module as "primary".
We had that for --fuzz-split mode, and this just expands that to any
time we have a second module.