[wasm-split] Set import namespace only when given in multi-split
Currently, in the case of two-way split, we set `config.importNamespace`
only when `--import-namespace` option is given, and otherwise just use
the default namespace "primary". But in the case of multi split, we set
it regardless of whether the option is given, resulting in the ""
namespace when no option is given.

The default import namespace for multi-split was "" when the multi-split
function was first added (#6943), so I'm not sure whether this
difference was intentional. In case it wasn't, I think it makes sense to
be consistent between two-way split and multi-split. So this in effect
changes the import namespace in multi split tests from "" to `primary`.

This also factors out the part that moves `WasmSplitOptions` into
`Config` to a function.