[emscripten] Remove -sWASM_BIGINT flag
This flag has been enabled by default in emscripten for a while now
so I don't think we need this here anymore.

The main user-visible effect of this change is that it makes the
`module.i64.const` JS method now simply take a single bigint rather than
a pair numbers.