[emscripten] Simply LTO flag setup. NFC
We have a setting (BYN_ENABLE_LTO) which controls if LTO is enabled, and
this is enabled for emscripten by default.

With emscripten we also want to add `-flto=thin` to the linker flags
(on other targets its not needed but in emscripten we need this so the
linker knows to use specific LTO system libraries)

I'm not sure if was intentional to use `-flto` and some places and
`-flto=thin` in others, but this change consistency does thing LTO.

Also, don't enable LTO by default in debug builds.