GlobalStructInference: Optimize gets of immutable global struct data (#8005)
A `struct.get` of a global, where the global creates a struct, can be
optimized by other passes, but not in all cases: Precompute can only
handle Literals (as it uses the interpreter), and GUFA works on a type-
based manner. The general case is important too, which looks like this:

```wat
;; Global struct, immutable.
(global $vtable (ref $vtable)
  (struct.new $vtable
    (global.get $imported)
  )
)

;; Code reads the global, then a field. This can be a read of $imported.
(struct.get $vtable 0
  (global.get $vtable)
)
```

This PR adds it to GlobalStructInference. That pass has a bunch of
useful infrastructure for it (parsing of struct.news, unnesting of
values, etc.), making it simple there.

Fixes #8002