Fuzzer: Use existing mechanism to fix up non-nullable global exnrefs (#8021)
We can't easily make a non-nullable exnref, and use a throw in a block
to manufacture one when necessary. That doesn't work in the global
scope. When in that bad situation, emit a RefAsNonNull to "fix" it,
which is not valid (like a throw+block are not valid), but existing code
for global generation notes that and fixes it up.