Fuzzer: Merge and optimize even with closed world in Two() (#7963)
The key idea here is that we can take two separate modules, merge them, and then
optimize in the most aggressive manner (even closed world) internally, as any
optimization change after the merge is a bug. (We must still avoid optimizing
the separate modules in closed world, which happens later in this fuzzer.)

Compare the full merged output to the pre-merged, for maximal coverage. This
requires fixing up export names after the merge, which is a little annoying, but
seems worth it.