[Custom Descriptors] Exact imports require CD
Validate that exact function imports require custom descriptors. An
alternative approach (that we may switch to in the future if necessary)
would be to always allow exact imports, but then strip them out in the
binary writer when custom descriptors are not enabled. For now take the
more explicit what-you-see-is-what-you-get approach.