Remove using microsoft STL extension in file.cpp
Standard std::[io]fstream does not have a constructor from `std::wstring`, but from `std::string` or `const wchar_t*`. Microsoft STL has an extension that supports a `std::wstring` constructor, which the current code depends on. However, that extension does not exist in libc++ and therefore fails when trying to use libc++ with clang-cl on Windows.

This patch uses the standard-compliant constructor and builds with libc++.