Costs: Adjust the cost of calls (#8047)
The cost of "4" earlier was in place because call costs don't really matter
(the effects prevent removing/replacing them), but there are calls without
effects, like with the proper intrinsic. Model call costs as very expensive.

The monomorphize test change is because it has hardcoded costs, and
calls are now too "heavy" for what it wants. The calls are replaced by
other code that does a bunch of work.

Fixes #8044