[Custom Descriptors] Use exact imports in wasm-split (#8043)
When custom descriptors are enabled, import primary functions into
secondary modules using exact imports. This allows references to those
functions in the secondary module to remain exact, just like they are
before splitting. Remove the exact casts we previously inserted to fix
this problem.

Do not use exact imports when custom descriptors are not enabled. In
principle this could cause validation errors because we allow e.g. exact
locals even when custom descriptors are not enabled. This could be
worked around in the future either by running a pass to remove exactness
before splitting or by always using and allowing exact imports, but then
emitting them as inexact imports when custom descriptors are disabled.