Fuzzer: Remove assert from Two() about not ignoring only after merge (#8049)
The logic assumed that a valid run that becomes ignored (host limit etc.) after
merge must be due to optimizations. But it turns out that isn't always so, as
merging makes calls go within a module, and the limits on the call stack are
a bit different, so with great luck we can notice a difference there. Remove the
incorrect assertion.