[Custom Descriptors] Add struct.new_desc
and struct.new_default_desc instructions that take descriptor operands.
These will replace the overload of struct.new and struct.new_default for
types that have descriptors. It will be invalid to use the new
instructions on types that do not have descriptors and to use the old
instructions on types that do have descriptors.

For now, to allow for a graceful transition, accept the old overloaded
uses of struct.new and struct.new_default where the new instructions
will eventually be required.