DeadArgumentElimination: Skip unprofitable single-call chains (#8072)
If a codebase has a long chain of single calls

a -> b -> c -> d

then we can end up in a very slow and unprofitable situation, removing
params from c's call to d, which then means c does not use some params,
and then we go back and so forth. Each step back requires a full scan of
the code. We could develop a more sophisticated IR to handle this, but it
would need to track that the local.get of the incoming param is only used
in an outgoing param, etc., which is not trivial, and these chains are
unprofitable in another way: single calls like this are inlined anyhow,
making this work redundant.

For simplicity, this PR detects the most obvious case of such a chain -
that an iteratiion of work only found a single single-caller call to remove
params from - and stops removing params from that point.

This makes this pass 30% faster on a large Dart testcase, which makes
-O3 9% faster. On other large wasm files I see much smaller benefits,
but it helps sometimes there too.

This is not truly NFC as while inlining will handle this case after, we do
alter the order a bit, leading to slightly different outputs sometimes. (The
changes are not better or worse, just noise.)