Keep instantiations alive on failure to prevent use-after-free when a partially-instantiated module writes a function reference to an imported table (#8111)
Fixes #8108. In linking0.wast, a module writes a function reference to
an imported memory but fails to instantiate completely due to an out of
bounds memory access. In this case, the function reference is expected
to stay alive even though instantiation didn't complete, but our code
currently drops the last reference in this case which causes a segfault.
Change the code to keep the reference alive even when instantiation
fails.

This is pessimistic since some modules may fail to instantiate without
writing any references to imported tables, but it avoids dangling
pointers in those cases that do write references.