Stop using operator[] only for default-insertion into maps (#8129)
Upstream libc++ recently added the [[nodiscard]] attribute to the return
value
of operator[] on maps, meaning the compiler now warns on uses of this
operator
that only want to insert rather than also use the result.

Switch to using try_emplace, which does not construct the value unless
insertion
succeeds, and is IMO more readable than operator[] for this purpose.

Fixes #8124