Fix assertion failure in unsubtyping (#8148)
When noting a subtype relationship, Unsubtyping asserts that if the
subtype is a tuple, then the supertype must be a tuple as well. This
assertion was violated when SubtypingDiscoverer noted a subtyping
between an unreachable br_if's tuple value and the br_if's type. Fix the
problem by not noting this subtyping for unreachable br_ifs.