[NFC] Inlining: Avoid wasteful non-nullable local fixups (#8147)
Mark the scanning and planning passes as not modifying the IR at all.

Mark the doInlining and Inlining passes as not needing non-nullable
fixups, as we apply those fixups when we inline (update the comment
there to clarify that).

This makes `--inlining` 8% faster on a Dart testcase I am looking at.
That is the result of avoiding 87% of the scans for non-nullable
fixups. The `--inlining-optimizing` pass, however, has a more modest
speedup (since it optimizes after each inlining anyhow, which is even
larger work): 10% fewer scans, and wall time is within noise. (Still,
this does avoid some wasted work even there, and marking sub-passes
that do not modify the IR as not doing so seems like a win for
readability.)