TypeRefining-GUFA: Fix bug with considering continuation fields empty (#8150)
We must skip such fields and not refine them, as we can't use casts to
fix up refinement issues. But we were just not writing anything to GUFA's
tracking of that location, which meant it looked unwritten - hence any
reads would be unreachable, and trap. To fix this, just write the
current type there.